Migrate user profile to Home
